Reduce copied code
