+++ performance/src/test/com/clickhouse/benchmark/clients/QueryClient.java
@@ -1,8 +1,6 @@
 package com.clickhouse.benchmark.clients;
 
-import com.clickhouse.client.ClickHouseClient;
 import com.clickhouse.client.ClickHouseResponse;
-import com.clickhouse.client.api.Client;
 import com.clickhouse.client.api.data_formats.ClickHouseBinaryFormatReader;
 import com.clickhouse.client.api.query.QueryResponse;
 import com.clickhouse.client.config.ClickHouseClientOption;
@@ -13,39 +11,16 @@ import org.openjdk.jmh.annotations.Level;
 import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
-import org.openjdk.jmh.annotations.TearDown;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @State(Scope.Benchmark)
 public class QueryClient extends BenchmarkBase {
     private static final Logger LOGGER = LoggerFactory.getLogger(QueryClient.class);
-    ClickHouseClient clientV1;
-    Client clientV2;
 
     @Setup(Level.Trial)
     public void setup(DataState dataState) throws Exception {
         super.setup(dataState, true);
-    }
-
-    @Setup(Level.Iteration)
-    public void setUpIteration() {
-        LOGGER.info("Setup Each Invocation");
-        clientV1 = getClientV1();
-        clientV2 = getClientV2();
-
-    }
-
-    @TearDown(Level.Iteration)
-    public void tearDownIteration() {
-        if (clientV1 != null) {
-            clientV1.close();
-            clientV1 = null;
-        }
-        if (clientV2 != null) {
-            clientV2.close();
-            clientV2 = null;
-        }
     }
 
 
@@ -53,7 +28,7 @@ public class QueryClient extends BenchmarkBase {
     public void queryV1(DataState dataState) {
         try {
             try (ClickHouseResponse response = clientV1.read(getServer())
-                    .query("SELECT * FROM `" + DB_NAME + "`.`" + dataState.dataSet.getTableName() + "`")
+                    .query("SELECT * FROM `" + DB_NAME + "`.`" + dataState.dataSet.getTableName() + "` LIMIT " + dataState.limit)
                     .format(ClickHouseFormat.RowBinaryWithNamesAndTypes)
                     .option(ClickHouseClientOption.ASYNC, false)
                     .executeAndWait()) {
@@ -71,7 +46,7 @@ public class QueryClient extends BenchmarkBase {
     @Benchmark
     public void queryV2(DataState dataState) {
         try {
-            try(QueryResponse response = clientV2.query("SELECT * FROM `" + dataState.dataSet.getTableName() + "`").get()) {
+            try(QueryResponse response = clientV2.query("SELECT * FROM `" + dataState.dataSet.getTableName() + "` LIMIT " + dataState.limit).get()) {
                 ClickHouseBinaryFormatReader reader = clientV2.newBinaryFormatReader(response);
                 while (reader.next() != null) {//Compiler optimization avoidance
                     for (int i = 1; i <= dataState.dataSet.getSchema().getColumns().size(); i++) {
